+++ modules/tinymce/src/core/main/ts/api/Settings.ts
@@ -5,7 +5,7 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Arr, Fun, Obj, Strings, Type } from '@ephox/katamari';
+import { Arr, Obj, Strings, Type } from '@ephox/katamari';
 
 import { UploadHandler } from '../file/Uploader';
 import DOMUtils from './dom/DOMUtils';
@@ -72,8 +72,6 @@ const shouldEndContainerOnEmptyBlock = (editor: Editor): boolean => editor.getPa
 const getFontStyleValues = (editor: Editor): string[] => Tools.explode(editor.getParam('font_size_style_values', 'xx-small,x-small,small,medium,large,x-large,xx-large'));
 
 const getFontSizeClasses = (editor: Editor): string[] => Tools.explode(editor.getParam('font_size_classes', ''));
-
-const getImagesDataImgFilter = (editor: Editor): (imgElm: HTMLImageElement) => boolean => editor.getParam('images_dataimg_filter', Fun.always, 'function');
 
 const isAutomaticUploadsEnabled = (editor: Editor): boolean => editor.getParam('automatic_uploads', true, 'boolean');
 
@@ -226,7 +224,6 @@ export {
   getFontSizeClasses,
   getIconPackName,
   getIconsUrl,
-  getImagesDataImgFilter,
   isAutomaticUploadsEnabled,
   shouldReuseFileName,
   shouldReplaceBlobUris,