+++ modules/tinymce/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Removed deprecated `Schema` settings #TINY-7821
+- Removed deprecated `file_browser_callback_types`, `force_hex_style_colors` and `images_dataimg_filter` settings #TINY-7823
 - The legacy `mobile` theme has been removed #TINY-7832
 
 ## 5.10.0 - 2021-10-11